+++ app/controllers/api/dashboards_controller.rb
@@ -8,6 +8,7 @@ class Api::DashboardsController < ApiController
   before_action :ensure_user_has_requested_apps, only: [:create, :update]
   before_action :ensure_is_manager_or_admin, only: :update
   before_action :ensure_is_admin_for_highlighting, only: [:create, :update]
+  before_action :ensure_is_admin_for_featuring, only: [:create, :update]
 
   def index
     if params.include?('user.role') && @user&.dig('role').eql?('ADMIN')
@@ -135,19 +136,27 @@ class Api::DashboardsController < ApiController
     render json: {errors: [{status: '403', title: 'You need to be an ADMIN to create/update the is-highlighted attribute of the dashboard'}]}, status: 403
   end
 
+  def ensure_is_admin_for_featuring
+    return true unless request.params.dig('data', 'attributes', 'is-featured').present?
+    return true if request.params.dig('data', 'attributes', 'is-featured').present? and @user[:role].eql? "ADMIN"
+    render json: {errors: [{status: '403', title: 'You need to be an ADMIN to create/update the is-featured attribute of the dashboard'}]}, status: 403
+  end
+
   def get_user
     @user = params['loggedUser'].present? ? JSON.parse(params['loggedUser']) : nil
   end
 
   def dashboard_params_get
-    params.permit(:name, :published, :private, :user, :application, 'is-highlighted'.to_sym, user: [], :filter => [:published, :private, :user])
+    params.permit(:name, :published, :private, :user, :application, 'is-highlighted'.to_sym,
+                  'is-featured'.to_sym, user: [], :filter => [:published, :private, :user])
   end
 
   def dashboard_params_create
     new_params = ActiveModelSerializers::Deserialization.jsonapi_parse(params)
     new_params = ActionController::Parameters.new(new_params)
     new_params.permit(:name, :description, :content, :published, :summary, :photo,
-                      :user_id, :private, :production, :preproduction, :staging, :is_highlighted, application:[])
+                      :user_id, :private, :production, :preproduction, :staging,
+                      :is_highlighted, :is_featured, application:[])
   rescue
     nil
   end
@@ -156,7 +165,8 @@ class Api::DashboardsController < ApiController
     new_params = ActiveModelSerializers::Deserialization.jsonapi_parse(params)
     new_params = ActionController::Parameters.new(new_params)
     new_params.permit(:name, :description, :content, :published, :summary,
-                      :photo, :private, :production, :preproduction, :staging, :is_highlighted, application:[])
+                      :photo, :private, :production, :preproduction, :staging,
+                      :is_highlighted, :is_featured, application:[])
   rescue
     nil
   end