+++ app/serializers/dashboard_serializer.rb
@@ -23,13 +23,14 @@
 #  staging            :boolean          default(FALSE)
 #  application        :string           default(["\"rw\""]), not null, is an Array
 #  is_highlighted     :boolean          default(FALSE)
+#  is_featured        :boolean          default(FALSE)
 #
 
 class DashboardSerializer < ActiveModel::Serializer
   attributes :id, :name, :slug, :summary, :description,
              :content, :published, :photo, :user_id, :private,
              :production, :preproduction, :staging, :user, :application,
-             :is_highlighted
+             :is_highlighted, :is_featured
 
   def photo
     {