+++ db/migrate/20191218122116_add_is_featured_field_to_dashboards.rb
@@ -1,0 +1,5 @@
+class AddIsFeaturedFieldToDashboards < ActiveRecord::Migration[5.1]
+  def change
+    add_column :dashboards, :is_featured, :boolean, default: false
+  end
+end